Writing Correction Price Services
1-Writing Correction Page
2-Add/Remove/Edit For Writing Correction Prices
3-Add Writing Types

+++ HadisIelts/Shared/Requests/Payment/ProcessWritingFilesRequest.cs
@@ -24,7 +24,7 @@ namespace HadisIelts.Shared.Requests.Payment
     {
         public WritingFile WritingFile { get; set; }
         public int WordCount { get; set; }
-        public decimal Price { get; set; }
+        public uint Price { get; set; }
 
     }
 }
